+++ src/app/agentConfigs/legacyAdapter.ts
